+++ src/de/crysxd/ownfx/SimpleProfileWrapper.java
@@ -4,10 +4,10 @@ package de.crysxd.ownfx;
 public class SimpleProfileWrapper {
 	
 	private String name;
-	private int id;
+	private long id;
 	private boolean active;
 	
-	public SimpleProfileWrapper(Profile p, int activeId) {
+	public SimpleProfileWrapper(Profile p, long activeId) {
 		this.name = p.getName();
 		this.id = p.getId();
 		this.active = this.id == activeId;
